fix namespaces of moved class OrderStatus in usages
